1. 整理文件夹命名； 2. QMUIHelper 增加 is58InchScreen 用于判断 iPhone X； 3. 修复 iOS 11 下在 QMUISearchController 的搜索结果界面里以 present 的方式打开新界面后返回到搜索结果界面，导航栏会错误地盖住搜索框； 4. QMUIToastView 适配键盘升起时的布局； 5. 优化 QMUIKit.framework 里自动更新 QMUIKit.h 的 Run Script 的写法； 6. 增加 removeFloatMin() 函数用于区分 CGFLOAT_MIN 和 0，同时修复 QMUIButton layoutSubviews 里由于拿 CGFLOAT_MIN 去计算，出现精度问题，导致布局不居中； 7. 将代码里的 iOS 版本判断写法全部换成新的 available() 判断语法，并去掉 iOS 7 的代码； 8. 修复 QMUITextView 在 scrollEnabled 为 NO 时快速换行，会 crash； 9. 配置表里与 navBar 相关的属性如果为 nil 则不要执行样式修改，否则会覆盖 storyboard 里的样式

+++ umbrellaHeaderFileCreator.py
@@ -1,4 +1,5 @@
-#!python
+#!/usr/bin/python
+#coding:utf-8
 
 import os
 
@@ -24,11 +25,20 @@ for filename in onlyfiles:
 
 ''' % (filename, filename)
 
-print 'umbrella creator: umbrellaFileContent = ' + umbrellaFileContent
+umbrellaFileContent = umbrellaFileContent.strip()
 
 f = open(umbrellaHeaderFilePath, 'r+')
 f.seek(0)
-f.write(umbrellaFileContent)
-f.truncate()
+oldFileContent = f.read().strip()
+if oldFileContent == umbrellaFileContent:
+  print 'umbrella creator: ' + umbrellaHeaderFileName + '的内容没有变化，不需要重写'
+else:
+  print 'umbrella creator: ' + umbrellaHeaderFileName + '的内容发生变化，开始重写'
+  print 'umbrella creator: umbrellaFileContent = ' + umbrellaFileContent
+
+  f.seek(0)
+  f.write(umbrellaFileContent)
+  f.truncate()
+
 f.close()
